+++ HISTORY.rst
@@ -9,6 +9,7 @@ Unreleased
 **Features and Improvements**
 * #173: Manual discount.
 * #161: Quantity is not editable if product has the 'no quantity' flag.
+* #174: Discount on specific product + and / or conditions for program
 
 **Bugfixes**